+++ src/main/java/com/juan/escuela/services/ProfesorService.java
@@ -1,8 +1,8 @@
 package com.juan.escuela.services;
 
 import com.juan.escuela.dto.ProfesorDto;
+import com.juan.escuela.exceptions.AppException;
 import com.juan.escuela.mappers.ProfesorMapper;
-import com.juan.escuela.models.Materia;
 import com.juan.escuela.models.Profesor;
 import com.juan.escuela.repositories.ProfesorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,16 +26,22 @@ public class ProfesorService {
     }
 
     public ProfesorDto getProfesorById(int id) {
-        Profesor profesor = profesorRepository.findById(id).orElse(null);
+        Profesor profesor = profesorRepository.findById(id)
+                .orElseThrow(() -> new AppException("el profesor con el id: " + id + " no existe", HttpStatus.BAD_REQUEST));
         return profesorMapper.toProfesorDto(profesor);
     }
 
-    public ProfesorDto updateProfesor(Profesor profesor) {
+    public ProfesorDto saveProfesor(Profesor profesor) {
+        System.out.println(profesor);
         Profesor profesorSave = profesorRepository.save(profesor);
         return profesorMapper.toProfesorDto(profesorSave);
     }
 
     public void deleteProfesorByid(int id){
+        if (!profesorRepository.existsById(id)) {
+            throw new AppException("el profesor con el id: " + id + " no existe", HttpStatus.BAD_REQUEST);
+        }
+
         profesorRepository.deleteById(id);
     }
